Create READ ME
*Overview
The point of this analysis was to ensure Steve’s parents got a good comparison of several green stocks for 2017 and 2018. 
They were originally interested in investing all their money into DAQO until they confronted Steve about it. Steve’s suggestion is to diversify and investigate other green stocks as well.
This analysis showed us the percentage returns for a myriad of other Stocks. This will ensure Steve’s parents will make the best decision regarding ROI.
Without previous data regarding the stocks, there would be nothing to compare and nothing to indicate which stocks were performing well and which stocks were not performing well. 
Stock analysis plays a big role in determining whether someone has the best likelihood of seeing a positive return on their investment or not based on their stock of interest. Our job is simply to create a visualization for Steve’s parents to see and determine whether they want to invest all their money into DQ or diversify

***Results
In 2017 the stock DQ performed well in the year 2017 with a return of 199.4%. In 2018 the DQ performed poorly with a negative return of 62.6%. 
We formatted the table through the macro Sub formatAllStocksAnalysisTable(). 
The code was designed to format the table and color the cells based on whether the return was positive or negative. 
This not only shows the ROI for both years but also makes the visualization clearer when presenting to a client. 
Green in the stock market is known as positive and red is known as negative. 
*Visuals: 
Adding =vbred and =vbgreen makes reading the chart easier for those unfamiliar with stocks. 
This will lead to better decision-making on the client’s part. We were able to see that a large portion of the stocks 
that performed well in 2017 did not perform well in 2018. 
Due to the major decrease, I think Steve’s parents will need to do more research to determine whether they think they should put all their money into DQ or not.
'Below is the segment of the code that is responsible for the bulk of the visualization. 

Worksheets("All Stocks Analysis").Activate
Range("A3:C3").Font.FontStyle = "Bold"
Range("A3:C3").Borders(xlEdgeBottom).LineStyle = xlContinuous
Range("B4:B15").NumberFormat = "#,##0"
Range("C4:C15").NumberFormat = "0.0%"
Columns("B").AutoFit
dataRowStart = 4
dataRowEnd = 15
For I = dataRowStart To dataRowEnd
             If Cells(I, 3) > 0 Then
            'Change cell color to green
            Cells(I, 3).Interior.Color = vbGreen
             ElseIf Cells(I, 3) < 0 Then
            'Change cell color to red
            Cells(I, 3).Interior.Color = vbRed
            Else
            'Clear the cell Color
            Cells(I, 3).Interior.Color = xlNone
            End If
The original code time to execute the year 2017 was 0.2734375 seconds and 0.2578125 seconds for 2018. 
The refactored version executed 2017 in 0.2578125 seconds and executed 2018 in 0.265625 seconds.

***Summary

The advantage of refactoring code, in general, is that it will save you time. 
You won’t be forced to go back and write an entirely new code which will in turn allow the client to see the finished product sooner. 
However, when refactoring code you always must ensure you have notes written to ensure you can tell what the code is doing. 
You don’t want to just plug in the sheet and the code output information that the client is not looking for. 
For the con, I will say you need to go back and spend a good amount of time reviewing the code and making sure it hits all the details that the 
client is asking for. Refactoring is more of a luxury that should be used when someone is asking for something similar to the work you've done in 
the past but is always modified to meet the current client’s needs. For VBA I’d say a pro is that it saves more time due to VBA being more of a 
specialized language. People are more likely to understand languages like Python and Java before VBA so not having to code the entire thing will 
lower some of the confusion on the developer's side. However, you will need to review all refactored code in detail before submitting it to ensure its accuracy. 
This is not a way to complete the assignment this is just a way to create the shell. The small details will need to be put in place by the developer.